refactor to decouple business logic
